update ursa final
