+++ vlmeval/vlm/ursa/__init__.py
@@ -1,2 +1,1 @@
 from .ursa_chat import UrsaChat
-from .ursa_model import deepseek_vl
