+++ vlmeval/vlm/ursa/ursa_model/clip_encoder.py
