+++ vlmeval/vlm/ursa/ursa_model/image_processing_vlm.py
